==> 03-cleaning.Rmd <==
# Data transformation


Since we have different data from both <https://datasets.imdbws.com>(prvided tsv) and <https://www.boxofficemojo.com/>(web scraping). There are some limits.

## Combining rating data from imdbdatasets

First of all we downloaded the title.basic.tsv.gz and title.rating.tsv.gz and read them by read_tsv()(the data are larger than the Storage limits in Github, Thus we didnot show the process of cleaning it here.)
the code is as follow
```
imdb_rank<-read_tsv("title.ratings.tsv")
head(imdb_rank)
imdb_basic<-read_tsv("title.basics.tsv")
head(imdb_basic)
imdb_movie<-imdb_basic%>%
filter(titleType=="movie")%>%
left_join(imdb_rank)%>%
na_if("\\N")%>%
select(-`endYear`,-titleType,-originalTitle,-isAdult)%>%
mutate(runtimeMinutes=as.integer(runtimeMinutes))%>%
separate(col=genres,into=c("g1","g2","g3"),sep = ",")
imdb_movie$startYear<-readr::parse_date(imdb_movie$startYear,format="%Y")
write_csv(imdb_movie,"data/imdb_movie.csv")
```

The cleaning imdb_movie data are in data folder.Here we only read the cleaned data.

```{r, include=FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE,echo=TRUE)
```

```{r}
library(tidyverse)
imdb_movie<-read_csv("data/imdb_movie.csv")
```

##  Cleaning boxoffice data

### Cleaning Top800 Boxoffice data

```{r}
library(tidyverse)
library(rvest)
library(robotstxt)

## Seperately read the movie information in each page.
paths_allowed("https://www.boxofficemojo.com/chart/ww_top_lifetime_gross/?area=XWW")
Yearly_boxoffice<-read_html("https://www.boxofficemojo.com/year/?ref_=bo_nb_di_secondarytab") %>% html_table()
Top200 <- read_html("https://www.boxofficemojo.com/chart/ww_top_lifetime_gross/?area=XWW") %>% html_table()
Top400<- read_html("https://www.boxofficemojo.com/chart/ww_top_lifetime_gross/?area=XWW&offset=200") %>% html_table()
Top600<-read_html("https://www.boxofficemojo.com/chart/ww_top_lifetime_gross/?area=XWW&offset=400") %>% html_table()
Top800<-read_html("https://www.boxofficemojo.com/chart/ww_top_lifetime_gross/?area=XWW&offset=600") %>% html_table()

#Combine them together
top800movie<-bind_rows(Top200,Top400)%>%bind_rows(Top600)%>%bind_rows(Top800)

#Combine the boxoffice information with basic information like title and rating
boxo<-left_join(top800movie,imdb_movie,by=c("Title" = "primaryTitle"))%>%
  group_by(Title)%>%
  slice_max(n = 1, order_by =numVotes)%>%
  arrange((Rank))%>%
  mutate(Total=as.numeric(gsub('[$,]','',`Worldwide Lifetime Gross`)),
         Domestic=as.numeric(gsub('[$,]','',`Domestic Lifetime Gross`)),
         Foreign=as.numeric(gsub('[$,]','',`Foreign Lifetime Gross`)))%>%
  select(-`Worldwide Lifetime Gross`,-`Foreign Lifetime Gross`,-`Domestic Lifetime Gross`,-Year)

#Save the data for following usuage.
write_csv(boxo,"data/imdb_Top800Movie_data.csv")
```

### Cleaning nearby 5 years Boxoffice data

```{r}
#nearby 5 years boxoffice data
start=2017
end=2020
boxall=read_html("https://www.boxofficemojo.com/year/world/2021/") %>% html_table()
for (i in start:end){
  url=paste("https://www.boxofficemojo.com/year/world/",i,"/",sep="")
  boxnow<-read_html(url) %>% html_table()
  boxall<-bind_rows(boxall,boxnow)}

# combine 5 years data with basic rating information.
box1721<-boxall%>%left_join(imdb_movie,by=c("Release Group" = "primaryTitle"))%>%
  group_by(`Release Group`)%>%
  slice_max(n = 1, order_by =numVotes)%>%
  select(-Rank)%>%
  mutate(Total=as.numeric(gsub('[$,]','',`Worldwide`)),
         Domestic=as.numeric(gsub('[$,]','',`Domestic`)),
         Foreign=as.numeric(gsub('[$,]','',`Foreign`)))%>%
  select(-`Worldwide`,-`Foreign`,-`Domestic`)%>%
  arrange(desc(Total))
write_csv(box1721,"data/imdb_20172021boxoffice_data.csv")
```

### Getting the yearly Boxoffice data.

```{r}
Yearly_boxoffice<-read_html("https://www.boxofficemojo.com/year/?ref_=bo_nb_di_secondarytab") %>% html_table()
Yearly_boxoffice<-Yearly_boxoffice%>%as.data.frame()%>%filter(Year!=2022)%>%mutate(Total=as.numeric(gsub('[$,()]','',`Total.Gross`)),Average=as.numeric(gsub('[$,]','',Average)))%>%select(-`Total.Gross`)
write_csv(Yearly_boxoffice,"data/imdb_yearlyboxoffice.csv")
```
