==> 06-interactive.Rmd <==
# Interactive component

## Finding the cluster and outlier in Parallel coordinal plot

```{r}
library(tidyverse)
library(parcoords)
library(lubridate)
box1721<-read_csv("data/imdb_20172021boxoffice_data.csv")
box1721_new<-box1721%>%drop_na(g1,startYear)%>%pivot_longer(
  cols = c("g1","g2","g3"),
  values_to = "genres",
  names_to = "type"
)
box1<-box1721_new%>%drop_na(genres,startYear)%>%
  group_by(startYear,genres)%>%
  summarise(averagegross=sum(Total)/n())%>%
  filter(startYear>ymd("2016-01-01"))%>%
  pivot_wider(
  values_from = averagegross,
  names_from = startYear
)
parcoords(box1, rownames = FALSE,
            brushMode = "1D-axes",
            alpha = 0.5,
            withD3 = TRUE,
            width = 900,
            height = 600)
```

## Top 10 rating movies under the influence of Covid-19

[Top10 rating movie information in 2019 and 2020](2019-2020Top10ratingdata.html){target="_blank"}


Explanation of the graph:

Here, we make an interactive graphic of D3 about the film industry 

 
1. From the perspective of top10rating films in the past two years, the overall range of ratings has increased and the highest score has increased, but the box office of Top10rating films has decreased. This may be due to that 2020, as the first year after the pandemic, Many films delay the release.

2. Although the highest-scoring films in both years were foreign films, 2019 was dominated by 《Parasite》,a famous film overworld, while 2020 was dominated by an Indian film which is less popular. This is probably due to the shutdown of the film industry in the film developed region.

3. The total number of voters for each film in 2020 dropped significantly. This has also led to the rise of Indian films outside Europe and America.
