==> 07-conclusion.Rmd <==
# Conclusion

Our analysis mainly focused on the total number of films released in the last 20 years and the box office data in the last five years. We analyzed from different angles the number of film type, film score, score, movie time, movie time variables for our concerns, also explored the interaction relationship between these variables, and some useful results are obtained, such as average highest box office, from science fiction film length is positively correlated with film score and so on.

We can use these results to predict the overall development of the film industry over the past 20 years and trends in the coming years. At the same time, it is worth noting that the novel coronavirus epidemic in 2020 has had a great impact on the film industry.

There are also some limitations in our project. For example, when analyzing the box office data, we only analyzed the representative movies of the top 200 tickets in each year, instead of analyzing all the movies. At the same time, we can also expand the variable category, such as the country of the film, main actors and so on. However, due to the limitations of our database, we are not able to complete these analyses at the moment, so these are also areas where we can improve. 

Our Future directions will focus on these limitations and expand the database more completely, so that we can conduct a more systematic and comprehensive analysis.
In this project, we have learned a lot of connections that we haven't found in ordinary times, such as the positive correlation between the number of movie reviewers and the length of the movie. In this project, we have learned that if we systematically analyze things in daily life, there will be some new discoveries.
